refactor: 주문 조회를 스토어로 분리
